feat:create erros responce

+++ src/services/announces/deleteAnnounce.service.ts
@@ -1,6 +1,7 @@
 import { Repository } from "typeorm";
 import { AppDataSource } from "../../data-source";
 import { Announce } from "../../entities/announce";
+import AppError from "../../errors/AppError";
 
 const deleteAnonnouceService = async (announceId: string) => {
   const announceRepository: Repository<Announce> =
@@ -9,6 +10,9 @@ const deleteAnonnouceService = async (announceId: string) => {
   const deletedAnnounce: any = await announceRepository.findOneBy({
     id: announceId,
   });
+  if (!deletedAnnounce) {
+    throw new AppError("announce not found", 404);
+  }
 
   await announceRepository.delete(deletedAnnounce);
 };
