+++ src/services/comment/deleteComment.service.ts
@@ -1,6 +1,7 @@
 import { Repository } from "typeorm";
 import { Comment } from "../../entities/comment";
 import { AppDataSource } from "../../data-source";
+import AppError from "../../errors/AppError";
 
 const deleteCommentService = async (commentId: string) => {
   const commentRepository: Repository<Comment> =
@@ -9,7 +10,10 @@ const deleteCommentService = async (commentId: string) => {
   const comment: any = await commentRepository.findOne({
     where: { id: commentId },
   });
-
+  
+  if (!comment) {
+    throw new AppError("comment not found", 404);
+  }
   await commentRepository.remove(comment);
 };
